feat(sdk-analytics): Refactor analytics window keys and enhance global interface
