+++ core-web/libs/sdk/analytics/src/lib/core/shared/models/index.ts
@@ -2,8 +2,6 @@
  * Central export point for all DotCMS Analytics models
  */
 
-import { ANALYTICS_WINDOWS_ACTIVE_KEY, ANALYTICS_WINDOWS_CLEANUP_KEY } from '../constants';
-
 // Data models (Analytics data structures)
 export * from './data.model';
 
@@ -15,11 +13,3 @@ export * from './request.model';
 
 // Library models (Internal SDK structures)
 export * from './library.model';
-
-// Extend Window interface to include our custom properties
-declare global {
-    interface Window {
-        [ANALYTICS_WINDOWS_CLEANUP_KEY]?: () => void;
-        [ANALYTICS_WINDOWS_ACTIVE_KEY]?: boolean;
-    }
-}
